Update: 1.0 2024-06-27 12:01AM
Took a break on the Web Dev course to focus on
the project. Spent around 4 hours working on it.

Learning:
-From working on the project, my understanding
of HTML and CSS has gotten better.
-I'm also continuing my studies away from the
computer and course. I'm using a coding app
called MIMO. It's great so far. So far it helped me
remember important elements in HTML.

Lessons:
-It's easier to understand whats being taught
in the course when I try to implement it into my
own site.
    -I don't always have to stick to the course.
     It's ok to stray away from the course to work
     on my own thing.
-I really enjoy learning HTML and CSS

Ideas:
-Might start putting these notes into my website
 to show the way I was thinking during my journey.
-I want to start putting the course projects into
website to show what i've done.